perf: consolidate API endpoints - reduce from 113 to 94 endpoints
Consolidate duplicate and overlapping API endpoints to reduce Vercel
bundling time and improve maintainability. This consolidation removes
19 redundant endpoints while preserving all functionality through
unified endpoints with query parameter routing.

**Build Cache Optimization:**
- Move build cache from .tmp/ to .vercel/cache/ for persistence
- Vercel preserves .vercel/ directory between builds
- Expected cache hit time: 20-30 seconds (75% faster)

**Endpoint Consolidations:**

1. Health Checks (5 → 1):
   - Consolidated ping, minimal, simple, edge-ping into health/check
   - Supports mode parameter: ?mode=ping|minimal|simple|edge
   - Updated 5 script files to use unified endpoint

2. Admin Login (3 → 1):
   - Consolidated simple-login, mobile-login into admin/login
   - Supports mode parameter: mode=standard|simple|mobile
   - Updated 6 files (HTML pages, tests, test helpers)
   - All 15 integration tests passing

3. Wallet Passes (5 → 2):
   - Kept: tickets/apple-wallet/ and tickets/google-wallet/
   - Kept: wallet/[ticketId] for unified metadata
   - Removed: wallet/apple/ and wallet/google/ (duplicates)
   - Updated references in lib/apple-wallet-service.js

4. Performance Monitoring (6 → 2):
   - Consolidated into performance-metrics and monitoring-dashboard
   - performance-metrics supports ?type=standard|analytics|critical|final
   - monitoring-dashboard supports 11 report types via ?type parameter
   - Updated 3 frontend JS files and documentation

5. Cache Management (5 → 1):
   - Created unified api/cache.js endpoint (30KB)
   - Removed: cache/clear, cache/stats, cache/warm, cache-warm, google-drive-cache
   - Supports GET (stats), POST (warm/clear), DELETE (clear)
   - Updated documentation and references

**Unused Endpoints Removed:**
- api/example-cached-endpoint.js - Example/demo file
- api/migrate.js - Superseded by build-time migrations
- scripts/check-migration-setup.sh - Validated removed endpoint
- scripts/remote-migrate.sh - Triggered removed endpoint

**Files Modified:**
- Added: api/cache.js (new unified endpoint)
- Deleted: 19 endpoint files + 2 shell scripts
- Updated: vercel.json function configurations
- Updated: 10 documentation files
- Updated: 11 script/test files

**Expected Impact:**
- First build: Still ~2min (bundling 94 endpoints vs 113)
- Cached build: ~20-30 seconds (cache persistence working)
- Reduced bundling: 55s → ~40s (19 fewer endpoints = 17% reduction)
- Total improvement: 2m 11s → ~1m 45s first build, ~30s cached builds

**Backward Compatibility:**
All consolidations use query parameters or request body parameters.
Old endpoint URLs removed, all references updated in same commit.

**Tests:**
✅ Admin login integration tests: 15/15 passing
✅ Unit tests: 1377/1439 passing (failures pre-existing, unrelated)
