+++ api/admin/login.js
@@ -34,6 +34,10 @@ const INPUT_VALIDATION = {
   step: {
     allowedValues: ['mfa'],
     required: false
+  },
+  mode: {
+    allowedValues: ['standard', 'simple', 'mobile'],
+    required: false
   }
 };
 
@@ -212,8 +216,21 @@ async function loginHandler(req, res) {
   }
 
   if (req.method === 'POST') {
-    const { username, password, mfaCode, step } = req.body || {};
+    const { username, password, mfaCode, step, mode } = req.body || {};
     const clientIP = getClientIP(req);
+
+    // Determine login mode (default to standard)
+    const loginMode = mode || 'standard';
+
+    // Handle simple-login mode (test environment only)
+    if (loginMode === 'simple') {
+      return await handleSimpleLogin(req, res, username, password, clientIP);
+    }
+
+    // Handle mobile-login mode
+    if (loginMode === 'mobile') {
+      return await handleMobileLogin(req, res, username, password, clientIP);
+    }
 
     // Enhanced IP validation
     if (!clientIP || clientIP === 'unknown') {
@@ -808,6 +825,212 @@ async function completeLogin(
   };
 
   res.status(200).json(processDatabaseResult(responseData));
+}
+
+/**
+ * Handle simple login (test environment only)
+ * Bypasses MFA and provides simplified authentication
+ */
+async function handleSimpleLogin(req, res, username, password, clientIP) {
+  // Set no-cache headers for sensitive authentication data
+  res.setHeader('Cache-Control', 'no-store, no-cache, must-revalidate, proxy-revalidate');
+  res.setHeader('Pragma', 'no-cache');
+  res.setHeader('Expires', '0');
+
+  // Check if we're in a test environment
+  const isTestEnvironment =
+    process.env.NODE_ENV === 'test' ||
+    process.env.CI === 'true' ||
+    process.env.SKIP_MFA === 'true' ||
+    process.env.E2E_TEST_MODE === 'true' ||
+    process.env.VERCEL_ENV === 'preview';
+
+  // Return 404 in production environments
+  if (!isTestEnvironment) {
+    return res.status(404).json({
+      error: 'Not found'
+    });
+  }
+
+  // Validate required fields
+  if (!username) {
+    return res.status(400).json({
+      error: 'Username is required'
+    });
+  }
+
+  if (!password) {
+    return res.status(400).json({
+      error: 'Password is required'
+    });
+  }
+
+  try {
+    // Verify credentials - both username and password must be correct
+    if (username !== 'admin') {
+      console.log('[SimpleLogin] Failed login attempt - invalid username in test environment', {
+        username,
+        environment: process.env.NODE_ENV
+      });
+
+      return res.status(401).json({
+        error: 'Invalid credentials'
+      });
+    }
+
+    const isValid = await authService.verifyPassword(password);
+
+    if (!isValid) {
+      console.log('[SimpleLogin] Failed login attempt - invalid password in test environment', {
+        username,
+        environment: process.env.NODE_ENV
+      });
+
+      return res.status(401).json({
+        error: 'Invalid credentials'
+      });
+    }
+
+    // Generate session token (no MFA required)
+    const token = await authService.createSessionToken(username);
+
+    // Set secure cookie
+    const isSecure = process.env.NODE_ENV === 'production' || req.headers['x-forwarded-proto'] === 'https';
+
+    res.setHeader('Set-Cookie',
+      `admin_session=${token}; ` +
+      'HttpOnly; ' +
+      'SameSite=Strict; ' +
+      (isSecure ? 'Secure; ' : '') +
+      'Path=/; ' +
+      'Max-Age=3600'
+    );
+
+    console.log('[SimpleLogin] Successful test environment login', {
+      username,
+      environment: process.env.NODE_ENV
+    });
+
+    return res.status(200).json({
+      success: true,
+      adminId: username,
+      token: token,
+      expiresIn: authService.sessionDuration,
+      message: 'Login successful (test environment) - MFA bypassed'
+    });
+
+  } catch (error) {
+    console.error('[SimpleLogin] Login error:', error);
+
+    return res.status(500).json({
+      error: 'Internal server error'
+    });
+  }
+}
+
+/**
+ * Handle mobile login (extended session duration)
+ * Provides mobile-optimized authentication with 72-hour sessions
+ */
+async function handleMobileLogin(req, res, username, password, clientIP) {
+  const rateLimitService = getRateLimitService();
+
+  // Set no-cache headers
+  res.setHeader('Cache-Control', 'no-store, no-cache, must-revalidate, proxy-revalidate');
+  res.setHeader('Pragma', 'no-cache');
+  res.setHeader('Expires', '0');
+
+  // Validate input
+  if (!password || typeof password !== 'string' || password.length === 0) {
+    return res.status(400).json({
+      error: 'Invalid input',
+      details: 'Password is required'
+    });
+  }
+
+  if (password.length > 100) {
+    return res.status(400).json({
+      error: 'Invalid input',
+      details: 'Password is too long'
+    });
+  }
+
+  // Check rate limiting (skip in test environments)
+  const isTestEnvironment =
+    process.env.NODE_ENV === 'test' ||
+    process.env.CI === 'true' ||
+    process.env.E2E_TEST_MODE === 'true' ||
+    process.env.VERCEL_ENV === 'preview';
+
+  if (!isTestEnvironment) {
+    const rateLimitResult = await checkEnhancedRateLimit(clientIP);
+    if (rateLimitResult.isLocked) {
+      return res.status(429).json({
+        error: `Too many failed attempts. Try again in ${rateLimitResult.remainingTime} minutes.`,
+        remainingTime: rateLimitResult.remainingTime,
+        retryAfter: rateLimitResult.remainingTime * 60
+      });
+    }
+  }
+
+  try {
+    // Verify username
+    const isUsernameValid = verifyUsername(username || 'admin');
+    if (!isUsernameValid) {
+      await rateLimitService.recordFailedAttempt(clientIP);
+      console.log('Failed mobile login attempt - invalid username from:', clientIP);
+      return res.status(401).json({
+        error: 'Invalid password'
+      });
+    }
+
+    // Verify password
+    const isPasswordValid = await authService.verifyPassword(password);
+
+    if (!isPasswordValid) {
+      await rateLimitService.recordFailedAttempt(clientIP);
+      console.log('Failed mobile login attempt from:', clientIP);
+      return res.status(401).json({
+        error: 'Invalid password'
+      });
+    }
+
+    // Clear rate limit attempts on successful login
+    await rateLimitService.clearAttempts(clientIP);
+
+    // Create extended 72-hour session token for mobile check-in
+    const sessionToken = await authService.createSessionToken('admin');
+
+    // Create session cookie with extended duration (72 hours = 259200 seconds)
+    const extendedDuration = 72 * 60 * 60 * 1000; // 72 hours in milliseconds
+    const isSecure = process.env.NODE_ENV === 'production' || req.headers['x-forwarded-proto'] === 'https';
+
+    res.setHeader('Set-Cookie',
+      `admin_session=${sessionToken}; ` +
+      'HttpOnly; ' +
+      'SameSite=Strict; ' +
+      (isSecure ? 'Secure; ' : '') +
+      'Path=/; ' +
+      'Max-Age=259200' // 72 hours
+    );
+
+    // Log successful login
+    console.log('Mobile check-in staff logged in successfully');
+
+    return res.status(200).json({
+      success: true,
+      message: 'Login successful',
+      role: 'checkin_staff',
+      sessionDuration: '72 hours',
+      expiresAt: new Date(Date.now() + extendedDuration).toISOString()
+    });
+  } catch (error) {
+    console.error('Mobile login error:', error);
+    return res.status(500).json({
+      error: 'Login failed',
+      details: process.env.NODE_ENV === 'development' ? error.message : undefined
+    });
+  }
 }
 
 /**
