--- api/health/edge-ping.js
@@ -1,16 +1,0 @@
-// Edge Runtime - bundles instantly, no Node.js overhead
-export const config = {
-  runtime: 'edge'
-};
-
-export default function handler(req) {
-  return new Response(JSON.stringify({
-    status: 'ok',
-    timestamp: Date.now()
-  }), {
-    status: 200,
-    headers: {
-      'content-type': 'application/json'
-    }
-  });
-}
