+++ docs/GOOGLE_DRIVE_INTEGRATION.md
@@ -70,11 +70,11 @@ Images are automatically categorized based on filename patterns:
 
 ### Google Drive Management
 - `GET /api/google-drive-health` - Service health and configuration status
-- `GET /api/google-drive-cache` - Cache status and metrics
-- `POST /api/google-drive-cache` - Warm up cache with fresh data (requires authentication)
-- `DELETE /api/google-drive-cache` - Clear cache (requires authentication)
+- `GET /api/cache?type=google-drive` - Cache status and metrics (requires admin authentication)
+- `POST /api/cache` - Warm up cache with fresh data (requires admin authentication, use `type=google-drive` and `action=warm`)
+- `DELETE /api/cache?type=google-drive` - Clear cache (requires admin authentication)
 
-**Note**: POST and DELETE operations on the cache management endpoint require an `X-API-Key` header with the `INTERNAL_API_KEY` value for security purposes.
+**Note**: All cache management operations require admin authentication via session token.
 
 ## Usage Examples
 
@@ -100,12 +100,38 @@ fetch('/api/google-drive-health')
   });
 ```
 
-### Clear Cache
+### Clear Google Drive Cache
 ```javascript
-fetch('/api/google-drive-cache', { method: 'DELETE' })
+fetch('/api/cache?type=google-drive', {
+  method: 'DELETE',
+  headers: {
+    'Authorization': 'Bearer <admin_token>'
+  }
+})
   .then(response => response.json())
   .then(data => {
     console.log('Cache cleared:', data.success);
+  });
+```
+
+### Warm Google Drive Cache
+```javascript
+fetch('/api/cache', {
+  method: 'POST',
+  headers: {
+    'Authorization': 'Bearer <admin_token>',
+    'Content-Type': 'application/json'
+  },
+  body: JSON.stringify({
+    type: 'google-drive',
+    action: 'warm',
+    year: 2026,
+    maxResults: 100
+  })
+})
+  .then(response => response.json())
+  .then(data => {
+    console.log('Cache warmed:', data.success);
   });
 ```
 
@@ -212,10 +238,18 @@ curl http://localhost:3000/api/google-drive-health
 
 ### Manual Cache Management
 
-Clear cache manually to force fresh data:
+Clear Google Drive cache manually to force fresh data:
 
 ```bash
-curl -X DELETE http://localhost:3000/api/google-drive-cache
+curl -X DELETE http://localhost:3000/api/cache?type=google-drive \
+  -H "Authorization: Bearer <admin_token>"
+```
+
+Check Google Drive cache status:
+
+```bash
+curl http://localhost:3000/api/cache?type=google-drive \
+  -H "Authorization: Bearer <admin_token>"
 ```
 
 ## Integration with Frontend
