+++ lib/apple-wallet-service.js
@@ -327,7 +327,7 @@ export class AppleWalletService {
 
     // Web service for updates (optional)
     if (process.env.WALLET_ENABLE_UPDATES === "true") {
-      pass.webServiceURL = `${this.baseUrl}/api/wallet/apple`;
+      pass.webServiceURL = `${this.baseUrl}/api/tickets/apple-wallet`;
       pass.authenticationToken = this.generateAuthToken(ticket.ticket_id);
     }
