+++ tests/integration/handler-test-helper.js
@@ -308,7 +308,6 @@ export async function testRequest(method, path, data = null, headers = {}) {
   // Map URL paths to API handler files
   const pathMappings = {
     '/api/admin/login': 'api/admin/login',
-    '/api/admin/simple-login': 'api/admin/simple-login',
     '/api/admin/verify-session': 'api/admin/verify-session',
     '/api/admin/dashboard': 'api/admin/dashboard',
     '/api/admin/registrations': 'api/admin/registrations',
